% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_movMF.R
\name{.posterior_from_ids}
\alias{.posterior_from_ids}
\title{Convert class IDs to posterior probability matrix}
\usage{
.posterior_from_ids(ids, k)
}
\arguments{
\item{ids}{Vector of class IDs}

\item{k}{Number of classes}
}
\value{
Binary posterior probability matrix (n x k)
}
\description{
Convert class IDs to posterior probability matrix
}
\keyword{internal}
