% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kappa_update}
\alias{kappa_update}
\title{Compute MLE update for kappa (concentration parameter)}
\usage{
kappa_update(x, method = "Newton_Fourier")
}
\arguments{
\item{x}{Data matrix (n x d) where each row is a unit vector on the hypersphere}

\item{method}{Method for computing kappa (default: "Newton_Fourier")}
}
\value{
Concentration parameter kappa
}
\description{
Compute MLE update for kappa (concentration parameter)
}
