% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.movMFnoise}
\alias{predict.movMFnoise}
\title{Predict method for movMFnoise objects}
\usage{
\method{predict}{movMFnoise}(
  object,
  newdata,
  what = c("dens", "cdens", "z", "map"),
  normalized = TRUE,
  logarithm = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{'movMFnoise'} object returned by \code{movMFnoise()} function
call.}

\item{newdata}{Optional new data matrix (n x d). If missing, training data are
used.}

\item{what}{A character string specifying what to retrieve: \code{"dens"}
returns a vector of values for the mixture density; \code{"cdens"} returns a
matrix of component densities for each mixture component (along the
columns); \code{"z"} returns a matrix of component posterior probabilities; \code{"map"}
returns the map classification.}

\item{logarithm}{A logical value indicating whether or not the logarithm of
the densities/probabilities should be returned.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Return predictions according to \code{what} argument.
}
\description{
Predict method for movMFnoise objects
}
