% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.movMFnoise}
\alias{print.summary.movMFnoise}
\title{Print method for summary.movMFnoise objects}
\usage{
\method{print}{summary.movMFnoise}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{A summary.movMFnoise object}

\item{digits}{Number of digits to print}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for summary.movMFnoise objects
}
