% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.movMFnoise}
\alias{summary.movMFnoise}
\title{Summary method for movMFnoise objects}
\usage{
\method{summary}{movMFnoise}(object, ...)
}
\arguments{
\item{object}{A movMFnoise object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for movMFnoise objects
}
