% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_movMF.R
\name{.compute_loglik}
\alias{.compute_loglik}
\title{Compute log-likelihood}
\usage{
.compute_loglik(x, theta, pro, d, Vinv = NULL)
}
\arguments{
\item{x}{Data matrix (n x d)}

\item{theta}{Canonical parameters (G x d)}

\item{pro}{Mixing proportions (length G or G+1 if noise)}

\item{d}{Number of dimensions}

\item{Vinv}{Inverse hypervolume for noise component (or NULL if no noise)}
}
\value{
Log-likelihood value
}
\description{
Compute log-likelihood
}
\keyword{internal}
