% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.movMFnoise}
\alias{predict.movMFnoise}
\title{Predict method for movMFnoise objects}
\usage{
\method{predict}{movMFnoise}(object, newdata, ...)
}
\arguments{
\item{object}{A movMFnoise object returned by em_movMF}

\item{newdata}{Optional new data matrix (n x d). If missing, uses training data}

\item{...}{Additional arguments (currently unused)}
}
\value{
List with components:
\item{classification}{Vector of component assignments}
\item{z}{Posterior probability matrix}
}
\description{
Predict method for movMFnoise objects
}
